requirements.txt
flask==2.2.3
werkzeug==2.2.3
wtforms==3.0.1
flask-wtf==1.1.1
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-login==0.6.2
flask-mail==0.9.1
flask-apscheduler==1.12.4
psycopg2-binary==2.9.5
gunicorn==20.1.0
python-dotenv==1.0.0
email-validator==2.0.0